+++ R/delete_files_plink.R
@@ -8,7 +8,7 @@
 #'
 #' @examples
 #' # create dummy BED/BIM/FAM files
-#' file <- 'delete-me-test' # no extension
+#' file <- tempfile('delete-me-test') # no extension
 #' # add each extension and create empty files
 #' file.create( paste0(file, '.bed') )
 #' file.create( paste0(file, '.bim') )